% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balls.R
\name{balls}
\alias{balls}
\title{Balls data}
\format{
A data.frame with 23 observations of failure times of bearing balls.
}
\source{
Lawless, J. F.  (2003). Statistical Models and Methods for Lifetime Data, Wiley,
Hoboken, NJ, USA.

Salam, S., Khan, Z., Ayed, H., Brahmia, A., Amin, A. (2021).
The Neutrosophic Lognormal Model in Lifetime Data Analysis: Properties and
Applications, \emph{Fuzzy Sets and Their Applications in Mathematics},
Article ID 6337759.
}
\description{
It is related to failure times of 23 bearing balls.
}
\examples{
data("balls")
balls
}
