% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbd.R
\name{NBD}
\alias{NBD}
\alias{pnbd}
\alias{dnbd}
\alias{qnbd}
\alias{rnbd}
\title{Neutrosophic Beta Distribution (NBD)}
\usage{
pnbd(x, alpha = 1, beta = 2)

dnbd(x, alpha = 1, beta = 2)

qnbd(q, alpha = 1, beta = 2)

rnbd(n, alpha = 1, beta = 2)
}
\arguments{
\item{x}{scaler or vector or matrix lower and upper of values at which the pdf or cdf needs to be computed.}

\item{alpha}{the value or vector lower and upper of the first shape parameter. must be positive.}

\item{beta}{the value or vector lower and upper of the second shape parameter, must be positive.}

\item{q}{scaler or vector of probabilities at which the quantile needs to be computed.}

\item{n}{number of random numbers to be generated.}
}
\value{
\code{pnbd} gives the distribution function,
 \code{dnbd} gives the density,
 \code{qnbd} gives the quantile function and
 \code{rnbd} generates random variables from the Neutrosophic Beta Distribution (NBD).
}
\description{
Computes the pdf, cdf, quantile and random numbers of the nuetrosophic generalized exponential distribution.
\deqn{f_N(X)=\frac{X^{\alpha_N-1}(1-X)^{\beta_N-1}}{\beta\left(\alpha_N, \beta_N\right)}}
for   \eqn{\alpha_N > 0}, the first shape parameter, and \eqn{\beta_N > 0}, the second shape parameter.
}
\examples{
x <- seq(0.1, 1, length.out = 21)
x2 <- matrix(seq(0.1, 2, length.out = 40), ncol = 2)
pnbd(x)
pnbd(x, alpha = 2, beta = 1)
pnbd(x2, alpha = c(1, 2), beta = c(2, 2))
dnbd(x, alpha = 1, beta = 2)
curve(dnbd, .1, 3)
dnbd(x2, alpha = c(1, 2), beta = c(2, 2))
qnbd(x, alpha = 1, beta = 2)
x2 <- matrix(seq(0.1, 1, length.out = 40), ncol = 2)
qnbd(x2, alpha = c(1, 2), beta = c(2, 2))
n <- 10
rnbd(n, alpha = 2, beta = 1)
rnbd(n, alpha = c(1, 2), beta = c(1, 1))
}
\references{
Sherwani, Rehan A. Khan, et al. Neutrosophic beta distribution with properties and applications. Infinite Study, 2021.
}
