% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbind.R
\name{NBIND}
\alias{NBIND}
\alias{dnbind}
\alias{pnbind}
\alias{qnbind}
\alias{rnbind}
\title{Neutrosophic Binomial Distribution (NBIND)}
\usage{
dnbind(x, size, prob)

pnbind(q, size, prob, lower.tail = TRUE)

qnbind(p, size, prob)

rnbind(n, size, prob)
}
\arguments{
\item{x}{a vector or matrix of observations for which the pdf needs to be computed.}

\item{size}{number of trials (zero or more), which must be a positive interval.}

\item{prob}{probability of success on each trial, \eqn{0 < prob <= 1}.}

\item{q}{a vector or matrix of quantiles for which the cdf needs to be computed.}

\item{lower.tail}{logical; if TRUE (default), probabilities are
\eqn{P(X \ge x)}; otherwise, \eqn{P(X >x)}.}

\item{p}{a vector or matrix of probabilities for which the quantile needs to be computed.}

\item{n}{number of random values to be generated.}
}
\value{
\code{pnbind} gives the distribution function,
 \code{dnbind} gives the density,
 \code{qnbind} gives the quantile function and
 \code{rnbind} generates random variables from the Binomial Poisson Distribution.
}
\description{
Density, distribution function, quantile function and random
generation for the nuetrosophic Binomial distribution with
parameters \eqn{n} and \eqn{p_N}.
}
\details{
The neutrosophic Binomial distribution with parameters \eqn{n} and \eqn{p_N}
has the density
\deqn{f_X(x)=\bigg(\begin{array}{c}n \\ x\end{array}\bigg) p_N^{x}\left(1-p_N\right)^{n-x}}
for \eqn{n \in \{1, 2, \ldots\}} and \eqn{p_N \in (p_L, p_U)} which must be \eqn{0<p_N<1}
and \eqn{x \in \{0, 1, 2, \ldots\}}.
}
\examples{
dnbind(x, size = 2, prob= 0.5)
dnbind(x2, lambda = c(2, 2))
x <- 1:10
x2 <- matrix(1:20, ncol = 2)
pnbind(x, size = 2, prob= 0.5)
pnbind(x2, size = c(2,2), prob = c(.3 ,.6))
q1 <- seq(0.1, 1, length.out = 40)
qnbind(q1, size = 2, prob= 0.5)
q2 <- matrix(seq(0.1, 1, length.out = 40), ncol = 2)
qnbind(q2, lambda = c(2, 2))
n <- 10
rnbind(n, size = 2, prob= 0.5)
rnbind(n, lambda = c(1, 2))
}
\references{
Granados, C. (2022).
       Some discrete neutrosophic distributions with neutrosophic parameters based on neutrosophic random variables.
        \emph{Hacettepe Journal of Mathematics and Statistics}, 51(5), 1442-1457.
}
