% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndunifd.R
\name{NDUNIFD}
\alias{NDUNIFD}
\alias{dndunifd}
\alias{qndunifd}
\alias{rndunifd}
\title{Neutrosophic Discrete Uniform Distribution (NDUNIFD)}
\usage{
dndunifd(x, min, max)

qndunifd(p, min, max)

rndunifd(n, min, max)
}
\arguments{
\item{x}{a vector or matrix of observations for which the pdf needs to be computed.}

\item{min}{lower limits of the distribution. Must be finite.}

\item{max}{upper limits of the distribution. Must be finite.}

\item{p}{a vector or matrix of probabilities for which the quantile needs to be computed.}

\item{n}{number of random values to be generated.}

\item{q}{a vector or matrix of quantiles for which the cdf needs to be computed.}

\item{lower.tail}{logical; if TRUE (default), probabilities are
\eqn{P(X \ge x)}; otherwise, \eqn{P(X >x)}.}
}
\value{
\code{pndunifd} gives the distribution function,
 \code{dndunifd} gives the density,
 \code{qndunifd} gives the quantile function and
 \code{rndunifd} generates random variables from the neutrosophic Discrete Uniform Distribution.
}
\description{
Density, distribution function, quantile function and random
generation for the nuetrosophic Discrete Uniform distribution with
parameters \eqn{a_N} and  \eqn{b_N}.
}
\details{
The neutrosophic Discrete Uniform distribution with parameters
\eqn{\max_N} and \eqn{\min_N} has the density
\deqn{f_N(x)=\frac{1}{b_N-a_N+1}}
for \eqn{a_N \in (a_L, a_U)}  lower parameter interval, \eqn{b_N \in (b_L,b_U)},
 upper parameter interval.
}
\examples{
dndunifd(x, min = 1, max = 2)

dndunifd(x2, min = c(1, 2), max = c(2, 2))
qndunifd(x, min = 1, max = 2)
x2 <- matrix(seq(0.1, 1, length.out = 40), ncol = 2)
qndunifd(x2, min = c(1, 2), max = c(2, 2))
n <- 10
rndunifd(n, min = 1, max = 2)
rndunifd(n, min = c(1, 2), max = c(1, 1))
}
\references{
Granados, C. (2022).
       Some discrete neutrosophic distributions with neutrosophic parameters based on neutrosophic random variables.
        \emph{Hacettepe Journal of Mathematics and Statistics}, 51(5), 1442-1457.
}
