% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ned.R
\name{NED}
\alias{NED}
\alias{dned}
\alias{pned}
\alias{qned}
\alias{rned}
\title{Neutrosophic Exponential Distribution (NED)}
\usage{
dned(x, theta)

pned(q, theta, lower.tail = TRUE)

qned(p, theta)

rned(n, theta)
}
\arguments{
\item{x}{a vector or matrix of observations for which the pdf needs to be computed.}

\item{theta}{the shape parameter, which must be a positive interval.}

\item{q}{a vector or matrix of quantiles for which the cdf needs to be computed.}

\item{lower.tail}{logical; if TRUE (default), probabilities are
\eqn{P(X \ge x)}; otherwise, \eqn{P(X >x)}.}

\item{p}{a vector or matrix of probabilities for which the quantile needs to be computed.}

\item{n}{number of random values to be generated.}
}
\value{
\code{pned} gives the distribution function,
 \code{dned} gives the density,
 \code{qned} gives the quantile function and
 \code{rned} generates random values from the neutrosophic exponential distribution.
}
\description{
Density, distribution function, quantile function and random
generation for the nuetrosophic exponential distribution with the
parameter rate \eqn{\theta_N}.
}
\details{
The neutrosophic exponential distribution with parameter \eqn{\theta_N}
has density
\deqn{f_N(x)=\theta_N \exp \left(-x \theta_N\right)}
for \eqn{x \ge 0} and \eqn{\theta_N \in (\theta_L, \theta_U)},
the rate parameter must be a positive interval and \eqn{x \ge 0}.
}
\examples{
# Example 4 of Duan et al. (2021)
data <- matrix(c(4,4,3.5,3.5,3.9,4.1,4.2,4.2,4.3,4.6, 4.7, 4.7), nrow = 6, ncol = 2, byrow = TRUE)
# Note that we have a matrix of nuetrosophic observations

# The density function of data with the estimated value for parameters based on Duan et al. (2021)
dned(data, theta = c(0.23, 0.24))
# The density function for the nuetrosophic observation (4,4.1)
dned(x = c(4,4.1), theta = c(0.23, 0.24))

# The density function for the nuetrosophic observation 4
# Here, 4 is equivalent to c(4,4).
dned(4, theta = c(0.23, 0.23))

# The cumulative distribution function for the nuetrosophic observation (4,4.1)
pned(c(4,4.1), theta = c(0.23, 0.24), lower.tail = TRUE)

pned(4, theta = c(0.23, 0.24))
# The first percentile
qned(p = 0.1, theta = 0.25)

# The quantiles
qned(p = c(0.25, 0.5, 0.75), theta = c(0.24, 0.25))

Simulate 10 numbers
rned(n = 10, theta = c(0.23, 0.24))
}
\references{
Duan, W., Q., Khan, Z., Gulistan, M., Khurshid, A. (2021). Neutrosophic
Exponential Distribution: Modeling and Applications for Complex Data Analysis,
\emph{Complexity}, 2021, 1-8.
}
