% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsgeom.R
\name{Neutrosophic Geometric}
\alias{Neutrosophic Geometric}
\alias{dnsgeom}
\alias{pnsgeom}
\alias{qnsgeom}
\alias{rnsgeom}
\title{Neutrosophic Geometric Distribution}
\usage{
dnsgeom(x, prob)

pnsgeom(q, prob, lower.tail = TRUE)

qnsgeom(p, prob)

rnsgeom(n, prob)
}
\arguments{
\item{x}{a vector or matrix of observations for which the pdf needs to be computed.}

\item{prob}{probability of success on each trial, \eqn{0 < prob <= 1}.}

\item{q}{a vector or matrix of quantiles for which the cdf needs to be computed.}

\item{lower.tail}{logical; if TRUE (default), probabilities are
\eqn{P(X \ge x)}; otherwise, \eqn{P(X >x)}.}

\item{p}{a vector or matrix of probabilities for which the quantile needs to be computed.}

\item{n}{number of random values to be generated.}
}
\value{
\code{pnsgeom} gives the distribution function,
 \code{dnsgeom} gives the density,
 \code{qnsgeom} gives the quantile function and
 \code{rnsgeom} generates random variables from the Geometric Distribution.
}
\description{
Density, distribution function, quantile function and random
generation for the nuetrosophic Geometric distribution with
parameter \code{prob} = \eqn{p_N}.
}
\details{
The neutrosophic Geometric distribution with parameter \eqn{p_N}
has the density
\deqn{f_X(x)=p_N\left(1-p_N\right)^x}
for \eqn{p_N \in (p_L, p_U)} which must be \eqn{0<p_N<1}
and \eqn{x \in \{0, 1, 2, \ldots\}}.
}
\examples{
dnsgeom(x, prob = 0.5)
dnsgeom(x2, lambda = c(2, 2))
x <- 1:10
x2 <- matrix(1:20, ncol = 2)
pnsgeom(x, prob = 0.5)
pnsgeom(x2, prob = c(.3, .6))
q1 <- seq(0.1, 1, length.out = 40)
qnsgeom(q1, prob = 0.5)
q2 <- matrix(seq(0.1, 1, length.out = 40), ncol = 2)
qnsgeom(q2, lambda = c(2, 2))
n <- 10
rnsgeom(n, prob = 0.5)
rnsgeom(n, lambda = c(1, 2))
}
\references{
Granados, C. (2022).
       Some discrete neutrosophic distributions with neutrosophic parameters based on neutrosophic random variables.
        \emph{Hacettepe Journal of Mathematics and Statistics}, 51(5), 1442-1457.
}
