% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nshyper.R
\name{Neutrosophic Hypergeometric}
\alias{Neutrosophic Hypergeometric}
\alias{dnshyper}
\alias{pnshyper}
\alias{qnshyper}
\alias{rnshyper}
\title{Neutrosophic Hypergeometric Distribution}
\usage{
dnshyper(x, N, K, k)

pnshyper(q, N, K, k, lower.tail = TRUE)

qnshyper(p, N, K, k)

rnshyper(nn, m, n, k)
}
\arguments{
\item{x}{a vector or matrix of observations for which the pdf needs to be computed.}

\item{N}{number of population size, which must be a positive interval.}

\item{K}{number of success states in the population, which must be a positive interval.}

\item{k}{number of draws (i.e. quantity drawn in each trial), , which must be a positive interval.}

\item{q}{a vector or matrix of quantiles for which the cdf needs to be computed.}

\item{lower.tail}{logical; if TRUE (default), probabilities are
\eqn{P(X \ge x)}; otherwise, \eqn{P(X >x)}.}

\item{p}{a vector or matrix of probabilities for which the quantile needs to be computed.}

\item{nn}{number of random values to be generated.}
}
\value{
\code{pnshyper} gives the distribution function,
 \code{dnshyper} gives the density,
 \code{qnshyper} gives the quantile function and
 \code{rnshyper} generates random variables from the HyperGeometric Distribution.
}
\description{
Density, distribution function, quantile function and random
generation for the nuetrosophic hypergeometric distribution with
parameters \eqn{m_N}, \eqn{n_N}, and \eqn{k_N}.
}
\details{
The neutrosophic hypergeometric distribution with parameters \eqn{k_N}
and \eqn{n_N} and \eqn{m_N}  has the density
\deqn{\frac{\left(\begin{array}{c}
m_N \\
x
\end{array}\right)\left(\begin{array}{c}
                       n_N \\
                       k_N-x
                       \end{array}\right)}{\left(\begin{array}{c}
                                                 m_N + n_N \\
                                                 k_N
                                                 \end{array}\right)}}
for \eqn{n_N \in (n_L, n_U)} which must be a positive interval and
 \eqn{k_N \in (k_L, k_U)} which must be a positive interval and
 \eqn{k \in \{0, 1, 2, \ldots\}} and
and \eqn{x \in \{0, 1, 2, \ldots, k\}}.
}
\examples{
dnshyper(x, N, K, k)
dnshyper(x2, N, K, k)
x <- 1:10
x2 <- matrix(1:20, ncol = 2)
pnshyper(x, N, K, k)
pnshyper(x2, N, K, k)
q1 <- seq(0.1, 1, length.out = 40)
qnshyper(q1, N, K, k)
q2 <- matrix(seq(0.1, 1, length.out = 40), ncol = 2)
qnshyper(q2, N, K, k)
n <- 10
rnshyper(n, N, K, k)
rnshyper(n, N, K, k)
}
\references{
Granados, C. (2022).
       Some discrete neutrosophic distributions with neutrosophic parameters based on neutrosophic random variables.
        \emph{Hacettepe Journal of Mathematics and Statistics}, 51(5), 1442-1457.
}
