% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nskumaraswamy.R
\name{Neutrosophic Kumaraswamy}
\alias{Neutrosophic Kumaraswamy}
\alias{dnskumaraswamy}
\alias{pnskumaraswamy}
\alias{qnskumaraswamy}
\alias{rnskumaraswamy}
\title{Neutrosophic Kumaraswamy Distribution}
\usage{
dnskumaraswamy(x, shape1, shape2)

pnskumaraswamy(q, shape1, shape2, lower.tail = TRUE)

qnskumaraswamy(p, shape1, shape2)

rnskumaraswamy(n, shape1, shape2)
}
\arguments{
\item{x}{a vector or matrix of observations for which the pdf needs to be computed.}

\item{shape1}{the shape parameter, which must be a positive interval.}

\item{shape2}{the shape parameter, which must be a positive interval.}

\item{q}{a vector or matrix of quantiles for which the cdf needs to be computed.}

\item{lower.tail}{logical; if TRUE (default), probabilities are
\eqn{P(X \ge x)}; otherwise, \eqn{P(X >x)}.}

\item{p}{a vector or matrix of probabilities for which the quantile needs to be computed.}

\item{n}{number of random values to be genelambdad.}
}
\value{
\code{pnskumaraswamy} gives the distribution function,
 \code{dnskumaraswamy} gives the density,
 \code{qnskumaraswamy} gives the quantile function and
 \code{rnskumaraswamy} genelambdas random values from the neutrosophic Kumaraswamy distribution.
}
\description{
Density, distribution function, quantile function and random
generation for the neutrosophic Kumaraswamy distribution with the
parameter \code{lambda} =  \eqn{\shape1_N}.
}
\details{
The neutrosophic Kumaraswamy distribution with parameter \eqn{\shape1_N}
has density
\deqn{f_N(x)=\shape1_N \exp \left(-x \shape1_N\right)}
for \eqn{x \ge 0} and \eqn{\shape1_N \in (\shape1_L, \shape1_U)},
the lambda parameter must be a positive interval and \eqn{x \ge 0}.
}
\examples{
# Example 4 of Duan et al. (2021)
data <- matrix(c(4, 4, 3.5, 3.5, 3.9, 4.1, 4.2, 4.2, 4.3, 4.6, 4.7, 4.7),
  nrow = 6, ncol = 2, byrow = TRUE
)

dnskumaraswamy(x = c(4, 4.1), shape1 = c(0.23, 0.24), shape2 = c(1, 2))
dnskumaraswamy(4, shape1 = c(0.23, 0.24), shape2 = c(1, 2))

# The cumulative distribution function for the nuetrosophic observation (4,4.1)
pnskumaraswamy(q = c(4, 4.1), shape1 = c(0.23, 0.24), shape2 = c(1, 2))
# The first percentile
qnskumaraswamy(p = 0.1, shape1 = 0.24, shape2 = 2)

# The quantiles
qnskumaraswamy(p = c(0.25, 0.5, 0.75), shape1 = c(0.23, 0.24), shape2 = c(1, 2))

# Simulate 10 numbers
rnskumaraswamy(n = 10, shape1 = c(0.23, 0.24), shape2 = c(1, 2))
}
\references{
Duan, W., Q., Khan, Z., Gulistan, M., Khurshid, A. (2021). Neutrosophic
Exponential Distribution: Modeling and Applications for Complex Data Analysis,
\emph{Complexity}, 2021, 1-8.
}
