% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nslaplace.R
\name{Neutrosophic Laplace}
\alias{Neutrosophic Laplace}
\alias{dnslaplace}
\alias{pnslaplace}
\alias{qnslaplace}
\alias{rnslaplace}
\title{Neutrosophic Laplace Distribution}
\usage{
dnslaplace(x, location, scale)

pnslaplace(q, location, scale, lower.tail = TRUE)

qnslaplace(p, location, scale)

rnslaplace(n, location, scale)
}
\arguments{
\item{x}{a vector or matrix of observations for which the pdf needs to be computed.}

\item{location}{the location parameter, which is the mean.}

\item{scale}{the scale parameter, Must consist of positive values.}

\item{q}{a vector or matrix of quantiles for which the cdf needs to be computed.}

\item{lower.tail}{logical; if TRUE (default), probabilities are
\eqn{P(X \ge x)}; otherwise, \eqn{P(X >x)}.}

\item{p}{a vector or matrix of probabilities for which the quantile needs to be computed.}

\item{n}{number of random values to be genelambdad.}
}
\value{
\code{pnslaplace} gives the distribution function,
 \code{dnslaplace} gives the density,
 \code{qnslaplace} gives the quantile function and
 \code{rnslaplace} genelambdas random values from the neutrosophic Laplace distribution.
}
\description{
Density, distribution function, quantile function and random
generation for the neutrosophic exponential distribution with the
parameter \code{lambda} =  \eqn{\location_N}.
}
\details{
The neutrosophic exponential distribution with parameter \eqn{\location_N}
has density
\deqn{f_N(x)=\location_N \exp \left(-x \location_N\right)}
for \eqn{x \ge 0} and \eqn{\location_N \in (\location_L, \location_U)},
the lambda parameter must be a positive interval and \eqn{x \ge 0}.
}
\examples{
# Example 4 of Duan et al. (2021)
data <- matrix(c(4, 4, 3.5, 3.5, 3.9, 4.1, 4.2, 4.2, 4.3, 4.6, 4.7, 4.7),
  nrow = 6, ncol = 2, byrow = TRUE
)

dnslaplace(x = c(4, 4.1), location = c(0.23, 0.24), scale = c(1, 2))
dnslaplace(4, location = c(0.23, 0.24), scale = c(1, 2))

# The cumulative distribution function for the nuetrosophic observation (4,4.1)
pnslaplace(q = c(4, 4.1), location = c(0.23, 0.24), scale = c(1, 2))
# The first percentile
qnslaplace(p = 0.1, location = 0.24, scale = 2)

# The quantiles
qnslaplace(p = c(0.25, 0.5, 0.75), location = c(0.23, 0.24), scale = c(1, 2))

# Simulate 10 numbers
rnslaplace(n = 10, location = c(0.23, 0.24), scale = c(1, 2))
}
\references{
Duan, W., Q., Khan, Z., Gulistan, M., Khurshid, A. (2021). Neutrosophic
Exponential Distribution: Modeling and Applications for Complex Data Analysis,
\emph{Complexity}, 2021, 1-8.
}
