% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsrayleigh.R
\name{Neutrosophic Rayleigh}
\alias{Neutrosophic Rayleigh}
\alias{dnsrayleigh}
\alias{pnsrayleigh}
\alias{qnsrayleigh}
\alias{rnsrayleigh}
\title{Neutrosophic Rayleigh Distribution}
\usage{
dnsrayleigh(x, theta)

pnsrayleigh(q, theta, lower.tail = TRUE)

qnsrayleigh(p, theta)

rnsrayleigh(n, theta)
}
\arguments{
\item{x}{a vector or matrix of observations for which the pdf needs to be computed.}

\item{theta}{the shape parameter, which must be a positive interval.}

\item{q}{a vector or matrix of quantiles for which the cdf needs to be computed.}

\item{lower.tail}{logical; if TRUE (default), probabilities are
\eqn{P(X \ge x)}; otherwise, \eqn{P(X >x)}.}

\item{p}{a vector or matrix of probabilities for which the quantile needs to be computed.}

\item{n}{number of random values to be generated.}
}
\value{
\code{dnsrayleigh} gives the density,
 \code{pnsrayleigh} gives the distribution function,
 \code{qnsrayleigh} gives the quantile function and
 \code{rnsrayleigh} generates random variables from the Neutrosophic Rayleigh Distribution.
}
\description{
Density, distribution function, quantile function and random
generation for the neutrosophic Rayleigh distribution with
parameter \eqn{\theta_N}.
}
\details{
The neutrosophic Rayleigh distribution with parameter \eqn{\theta}
has the density
\deqn{f_N(x)=\frac{x}{\theta_N^2} e^{-\frac{1}{2}\left(\frac{x}{\theta_N}\right)^2}}
for  \eqn{\theta_N \in (\theta_L, \theta_U)}, which must be a positive
interval and \eqn{x \ge 0}.
}
\examples{
data(remission)
dnsrayleigh(x = remission, theta = c(9.6432, 9.8702))

pnsrayleigh(q = 20, theta = c(9.6432, 9.8702))

# Calculate quantiles
qnsrayleigh(p = c(0.25, 0.5, 0.75), theta = c(9.6432, 9.8702))

# Simulate 10 values
rnsrayleigh(n = 10, theta = c(9.6432, 9.8702))

}
\references{
Khan, Z., Gulistan, M., Kausar, N. and Park, C. (2021).
Neutrosophic Rayleigh Model With Some Basic Characteristics and
Engineering Applications, in \emph{IEEE Access}, 9, 71277-71283.
}
