% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngd.R
\name{NGD}
\alias{NGD}
\alias{pngd}
\alias{dngd}
\alias{qngd}
\alias{rngd}
\title{Neutrosophic Gamma Distribution (NGD)}
\usage{
pngd(x, p = 1, lambda = 2)

dngd(x, p = 1, lambda = 2)

qngd(q, p = 1, lambda = 2)

rngd(n, p = 1, lambda = 2)
}
\arguments{
\item{x}{scaler or vector or matrix lower and upper of values at which the pdf or cdf needs to be computed.}

\item{p}{the value or vector lower and upper of the first shape parameter, must be positive.}

\item{lambda}{the value or vector lower and upper of the second scale parameter, must be positive.}

\item{q}{scaler or vector of probabilities at which the quantile needs to be computed.}

\item{n}{number of random numbers to be generated.}
}
\value{
\code{pngd} gives the distribution function,
 \code{dngd} gives the density,
 \code{qngd} gives the quantile function and
 \code{rngd} generates random variables from the Neutrosophic Gamma Distribution (NGD).
}
\description{
Computes the pdf, cdf, hdf, quantile and random numbers of the nuetrosophic generalized exponential distribution.
\deqn{f_n(z)=\frac{1}{\Gamma p_n \lambda_n^{p_n}} z^{p_n-1} e^{-\left(z / \lambda_n\right)}}
for   \eqn{p_n > 0}, the first shape parameter, and \eqn{\lambda_n > 0}, the second scale parameter.
}
\examples{
x <- seq(0.1, 1, length.out = 21)
x2 <- matrix(seq(0.1, 2, length.out = 40), ncol = 2)
pngd(x)
pngd(x, p = 2, lambda = 1)
pngd(x2, p = c(1, 2), lambda = c(2, 2))
dngd(x, p = 1, lambda = 2)
curve(dngd, .1, 3)
dngd(x2, p = c(1, 2), lambda = c(2, 2))
qngd(x, p = 1, lambda = 2)
x2 <- matrix(seq(0.1, 1, length.out = 40), ncol = 2)
qngd(x2, p = c(1, 2), lambda = c(2, 2))
n <- 10
rngd(n, p = 2, lambda = 1)
rngd(n, p = c(1, 2), lambda = c(1, 1))
}
\references{
Khan, Zahid, et al. "On statistical development of neutrosophic gamma distribution with applications to complex data analysis." Complexity 2021 (2021): 1-8.
}
