% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nge.R
\name{NGE}
\alias{NGE}
\alias{pnge}
\alias{dnge}
\alias{qnge}
\alias{rnge}
\alias{hnge}
\alias{snge}
\title{Neutrosophic Generalized Exponential (NGE)}
\usage{
pnge(x, v = 1, delta = 2)

dnge(x, v = 1, delta = 2)

qnge(q, v = 1, delta = 2)

rnge(n, v = 1, delta = 2)

hnge(x, v = 1, delta = 2)

snge(x, v = 1, delta = 2)
}
\arguments{
\item{x}{scaler or vector of values at which the pdf or cdf needs to be computed.}

\item{v}{the value of the first shape parameter.}

\item{delta}{the value of the second shape parameter, must be positive, the default is 2.}

\item{q}{scaler or vector of probabilities at which the quantile needs to be computed.}

\item{n}{number of random numbers to be generated.}
}
\value{
\code{pnge} gives the distribution function,
 \code{dnge} gives the density,
 \code{qnge} gives the quantile function,
 \code{hnge} gives the hazard function,
 \code{snge} gives the survival function and
 \code{rnge} generates random variables from the Neutrosophic Generalized Exponential (NGE).
}
\description{
Computes the pdf, cdf, hdf, quantile and random numbers of the
}
\examples{
x <- seq(0, 1, length.out = 21)
pnge(x)
pnge(x, v = 2, delta = 1)
dnge(x, v = 1, delta = 2)
curve(dnge, -3, 3)
qnge(x, v = 1, delta = 3)
n <- 10
rnge(n, v = 2, delta=1)
hnge(x, v = 1, delta = 3)
curve(hnge, 0, 3)
snge(x, v = 1, delta = 3)
curve(snge, 0, 3)
}
\references{
Rao, Gadde Srinivasa, Mina Norouzirad, and Danial Mazarei. "Neutrosophic Generalized Exponential Distribution with Application." Neutrosophic Sets and Systems 55.1 (2023): 28.
}
