% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nged.R
\name{NGED}
\alias{NGED}
\alias{dnged}
\alias{pnged}
\alias{qnged}
\alias{rnged}
\title{Neutrosophic Generalized Exponential Distribution (NGED)}
\usage{
dnged(x, nu, delta)

pnged(q, nu, delta, lower.tail = TRUE)

qnged(p, nu, delta)

rnged(n, nu, delta)
}
\arguments{
\item{x}{a vector or matrix of observations for which the pdf needs to be computed.}

\item{nu}{the scale parameter, which must be a positive interval.}

\item{delta}{the shape parameter, which must be a positive interval.}

\item{q}{a vector or matrix of quantiles for which the cdf needs to be computed.}

\item{lower.tail}{logical; if TRUE (default), probabilities are
\eqn{P(X \ge x)}; otherwise, \eqn{P(X >x)}.}

\item{p}{a vector or matrix of probabilities for which the quantile needs to be computed.}

\item{n}{number of random values to be generated.}
}
\value{
\code{pnged} gives the distribution function,
 \code{dnged} gives the density,
 \code{qnged} gives the quantile function and
 \code{rnged} generates random variables
 from the neutrosophic generalized exponential distribution.
}
\description{
Density, distribution function, quantile function and random
generation for the nuetrosophic generalized exponential
distribution with shape parameter\eqn{\delta_N} and scale parameter
\eqn{\nu_N}.
}
\details{
The neutrosophic generalized exponential distribution with parameters
\eqn{\delta} and \eqn{\nu} has density
\deqn{f_n(x)=\frac{\delta_N}{\nu_N}\left(1-\exp \left\{-\frac{x_N}{\nu_N}\right\}\right)^{\delta_N-1} e^{\left\{-\frac{x_N}{\nu_N}\right\}}}
for \eqn{\delta_N \in (\delta_L, \delta_U)}, the shape parameter
which must be a positive interval, and \eqn{\nu_N \in (\nu_L, \nu_U)}, the
scale parameter which must also be a positive interval, and \eqn{x \ge 0}.
}
\examples{
dnged(x = 1, nu = c(1, 2), delta = c(2, 2))

pnged(q=0.1, nu = c(1, 2), delta = c(2, 2))

x2 <- matrix(seq(0.1, 1, length.out = 40), ncol = 2)
pnged(x2, nu = c(1, 2), delta = c(2, 2))
qnged(x, nu = 1, delta = 2)

qnged(x2, nu = c(1, 2), delta = c(2, 2))
n <- 10
rnged(n = 10, nu = 2, delta = 1)
rnged(n = 10, nu = c(1, 2), delta = c(1, 1))
}
\references{
Rao, G. S., Norouzirad, M., and Mazarei . D. (2023). Neutrosophic
   Generalized Exponential Distribution with Application.
   \emph{Neutrosophic Sets and Systems}, 55, 471-485.
}
