% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhgeomd.R
\name{NHGEOMD}
\alias{NHGEOMD}
\alias{dnhgeomd}
\alias{pnhgeomd}
\alias{qnhgeomd}
\alias{rnhgeomd}
\title{Neutrosophic HyperGeometric Distribution (NHGEOMD)}
\usage{
dnhgeomd(x, N, K, k)

pnhgeomd(q, N, K, k, lower.tail = TRUE)

qnhgeomd(p, N, K, k)

rnhgeomd(n, N, K, k)
}
\arguments{
\item{x}{a vector or matrix of observations for which the pdf needs to be computed.}

\item{N}{number of population size, which must be a positive interval.}

\item{K}{number of success states in the population, which must be a positive interval.}

\item{k}{number of draws (i.e. quantity drawn in each trial), , which must be a positive interval.}

\item{q}{a vector or matrix of quantiles for which the cdf needs to be computed.}

\item{lower.tail}{logical; if TRUE (default), probabilities are
\eqn{P(X \ge x)}; otherwise, \eqn{P(X >x)}.}

\item{p}{a vector or matrix of probabilities for which the quantile needs to be computed.}

\item{n}{number of random values to be generated.}
}
\value{
\code{pnhgeomd} gives the distribution function,
 \code{dnhgeomd} gives the density,
 \code{qnhgeomd} gives the quantile function and
 \code{rnhgeomd} generates random variables from the HyperGeometric Distribution.
}
\description{
Density, distribution function, quantile function and random
generation for the nuetrosophic HyperGeometric distribution with
parameter \eqn{p_N}.
}
\details{
The neutrosophic HyperGeometric distribution with parameters \eqn{K_N}
and \eqn{N_N} and \eqn{k}
has the density
\deqn{\frac{\left(\begin{array}{c}
K_N \\
x
\end{array}\right)\left(\begin{array}{c}
                       N_N-K_N \\
                       k-x
                       \end{array}\right)}{\left(\begin{array}{c}
                                                 N_N \\
                                                 k
                                                 \end{array}\right)}}
for \eqn{N_N \in (N_L, N_U)} which must be a positive interval and
 \eqn{K_N \in (K_L, K_U)} which must be a positive interval and
 \eqn{k \in \{0, 1, 2, \ldots\}} and
and \eqn{x \in \{0, 1, 2, \ldots\}}.
}
\examples{
dnhgeomd(x, N, K, k)
dnhgeomd(x2, N, K, k)
x <- 1:10
x2 <- matrix(1:20, ncol = 2)
pnhgeomd(x, N, K, k)
pnhgeomd(x2, N, K, k)
q1 <- seq(0.1, 1, length.out = 40)
qnhgeomd(q1, N, K, k)
q2 <- matrix(seq(0.1, 1, length.out = 40), ncol = 2)
qnhgeomd(q2, N, K, k)
n <- 10
rnhgeomd(n, N, K, k)
rnhgeomd(n, N, K, k)
}
\references{
Granados, C. (2022).
       Some discrete neutrosophic distributions with neutrosophic parameters based on neutrosophic random variables.
        \emph{Hacettepe Journal of Mathematics and Statistics}, 51(5), 1442-1457.
}
