% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npd.R
\name{NPD}
\alias{NPD}
\alias{pnpd}
\alias{dnpd}
\alias{qnpd}
\alias{rnpd}
\title{Neutrosophic Poisson Distribution (NPD)}
\usage{
pnpd(p, lambda)

dnpd(x, lambda)

qnpd(q, lambda)

rnpd(n, lambda)
}
\arguments{
\item{p}{scaler or vector of probabilities at which the quantile needs to be computed.}

\item{lambda}{the value or vector lower and upper of (non-negative)the first shape parameter, must be positive.}

\item{x}{scaler or vector or matrix lower and upper of values at which
the pdf or cdf needs to be computed.}

\item{q}{scaler or vector of quantiles.}

\item{n}{number of random numbers to be generated.}
}
\value{
\code{pnpd} gives the distribution function,
 \code{dnpd} gives the density,
 \code{qnpd} gives the quantile function and
 \code{rnpd} generates random variables from the Neutrosophic Poisson Distribution (NPD).
}
\description{
Density, distribution function, quantile function and random generation for
the nuetrosophic Poisson distribution with parameter \eqn{\lambda_N}.
}
\details{
The neutrosophic Poisson distribution with parameter \eqn{\theta} has the density
\deqn{f_N(x)=e^{-\lambda_N}  \frac{\left(\lambda_N\right)^x}{x !}}
for   \eqn{\lambda_N > 0}.
}
\examples{
p <- 1:10
p2 <- matrix(1:20, ncol = 2)
pnpd(p, lambda = 1)
pnpd(p2, lambda = c(2, 3))
dnpd(x, lambda = 2)
dnpd(x2, lambda = c(2, 2))
q1 <- seq(0.1, 1, length.out = 40)
qnpd(q1, lambda = 2)
q2 <- matrix(seq(0.1, 1, length.out = 40), ncol = 2)
qnpd(q2, lambda = c(2, 2))
n <- 10
rnpd(n, lambda = 1)
rnpd(n, lambda = c(1, 2))
}
\references{
Alhabib, R., Ranna, M. M., Farah, H., Salama, A. A. (2018).
       Some neutrosophic probability distributions.
       \emph{Neutrosophic Sets and Systems},  22, 30-38.
}
