% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuifd.R
\name{NUNIFD}
\alias{NUNIFD}
\alias{dnunifd}
\alias{qnunifd}
\alias{rnunifd}
\title{Neutrosophic Uniform Distribution (NUNIFD)}
\usage{
dnunifd(x, min, max)

qnunifd(p, min, max)

rnunifd(n, min, max)
}
\arguments{
\item{x}{a vector or matrix of observations for which the pdf needs to be computed.}

\item{min}{lower limits of the distribution. Must be finite.}

\item{max}{upper limits of the distribution. Must be finite.}

\item{p}{a vector or matrix of probabilities for which the quantile needs to be computed.}

\item{n}{number of random values to be generated.}

\item{q}{a vector or matrix of quantiles for which the cdf needs to be computed.}

\item{lower.tail}{logical; if TRUE (default), probabilities are
\eqn{P(X \ge x)}; otherwise, \eqn{P(X >x)}.}
}
\value{
\code{pnunifd} gives the distribution function,
 \code{dnunifd} gives the density,
 \code{qnunifd} gives the quantile function and
 \code{rnunifd} generates random variables from the neutrosophic Uniform dDistribution.
}
\description{
Density, distribution function, quantile function and random
generation for the nuetrosophic Uniform distribution with
parameters \eqn{a_N} and  \eqn{b_N}.
}
\details{
The neutrosophic Rayleigh distribution with parameters
\eqn{\max_N} and \eqn{\min_N} has the density
\deqn{f_N(x)=\frac{1}{b_N-a_N}}
for \eqn{a_N \in (a_L, a_U)}  parameter interval, \eqn{b_N \in (b_L,b_U)},
 parameter interval, and \eqn{x > 0}.
}
\examples{
dnunifd(x, min = 1, max = 2)

dnunifd(x2, min = c(1, 2), max = c(2, 2))
qnunifd(x, min = 1, max = 2)
x2 <- matrix(seq(0.1, 1, length.out = 40), ncol = 2)
qnunifd(x2, min = c(1, 2), max = c(2, 2))
n <- 10
rnunifd(n, min = 1, max = 2)
rnunifd(n, min = c(1, 2), max = c(1, 1))
}
\references{
Alhasan, K. F. H. and Smarandache, F. (2019). Neutrosophic Uniform
   distribution and Neutrosophic Family Uniform Distribution,
   \emph{Neutrosophic Sets and Systems}, 28, 191-199.
}
