% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwd.R
\name{NWD}
\alias{NWD}
\alias{pnwd}
\alias{dnwd}
\alias{qnwd}
\alias{rnwd}
\title{Neutrosophic Weibull Distribution (NWD)}
\usage{
pnwd(x, beta = 1, alpha = 2)

dnwd(x, beta = 1, alpha = 2)

qnwd(q, beta = 1, alpha = 2)

rnwd(n, beta = 1, alpha = 2)
}
\arguments{
\item{x}{scaler or vector or matrix lower and upper of values at which the pdf or cdf needs to be computed.}

\item{beta}{the value or vector lower and upper of the first shape parameter. must be positive.}

\item{alpha}{the value or vector lower and upper of the second scale parameter, must be positive.}

\item{q}{scaler or vector of probabilities at which the quantile needs to be computed.}

\item{n}{number of random numbers to be generated.}
}
\value{
\code{pnwd} gives the distribution function,
 \code{dnwd} gives the density,
 \code{qnwd} gives the quantile function and
 \code{rnwd} generates random variables from the Neutrosophic Weibull Distribution (NWD).
}
\description{
Computes the pdf, cdf, hdf, quantile and random numbers of the nuetrosophic generalized exponential distribution.
\deqn{f_N(X)=\frac{\beta_N}{\alpha_N} X^{\beta_N-1} e^{-\left(X / \alpha_N\right)^{\beta_N}}}
for   \eqn{\beta_N > 0}, the first shape parameter, and \eqn{\alpha_N > 0}, the second scale parameter.
}
\examples{
x <- seq(0.1, 1, length.out = 21)
x2 <- matrix(seq(0.1, 2, length.out = 40), ncol = 2)
pnwd(x)
pnwd(x, beta = 1, alpha = 2)
pnwd(x2, beta = c(1, 2), alpha = c(2, 2))
dnwd(x, beta = 1, alpha = 2)
curve(dnwd, .1, 3)
dnwd(x2, beta = c(1, 2), alpha = c(2, 2))
qnwd(x, beta = 1, alpha = 2)
x2 <- matrix(seq(0.1, 1, length.out = 40), ncol = 2)
qnwd(x2, beta = c(1, 2), alpha = c(2, 2))
n <- 10
rnwd(n, beta = 1, alpha = 2)
rnwd(n, beta = c(1, 2), alpha = c(1, 1))
}
\references{
Sherwani, Rehan A. Khan, et al. Neutrosophic beta distribution with properties and applications. Infinite Study, 2021.
}
