% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remission.R
\name{remission}
\alias{remission}
\title{Remission data}
\format{
A data.frame with 128 observations of remission time in months of cancer patients.
}
\source{
Lee, E.T. and Wang, J. (2003), Statistical Methods for Survival Data Analysis.
 Vol. 476, John Wiley & Sons,Hoboken, NJ, USA.

Rao, G. S., Norouzirad, M., and Mazarei . D. (2023). Neutrosophic
   Generalized Exponential Distribution with Application.
   \emph{Neutrosophic Sets and Systems}, 55, 471-485.
}
\description{
It is related to remission time in months of 128 cancer patients.
}
\examples{
data("remission")
remission
}
