% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbd.R
\name{NBD}
\alias{NBD}
\alias{pnbd}
\alias{dnbd}
\alias{qnbd}
\alias{rnbd}
\title{Neutrosophic Beta Distribution (NBD)}
\usage{
pnbd(q, alpha, beta)

dnbd(x, alpha, beta)

qnbd(p, alpha, beta)

rnbd(n, alpha, beta)
}
\arguments{
\item{alpha}{the positive value or vector of the first shape parameter.}

\item{beta}{the positive value or vector of the second shape parameter.}

\item{x, q}{vector or matrix  of quantiles at which the pdf or cdf needs to be computed.}

\item{p}{vector or matrix of probabilities at which the quantile needs to be computed.}

\item{n}{number of random numbers to be generated.}
}
\value{
\code{pnbd} gives the distribution function,
\code{dnbd} gives the density,
\code{qnbd} gives the quantile function and
\code{rnbd} generates random variables from the neutrosophic Beta distribution.
}
\description{
Density, distribution function, quantile function and random generation for
the nuetrosophic Beta distribution with parameters \code{shape1}=\eqn{\alpha_N} and
\code{shape2}=\eqn{\beta_N}.
}
\details{
The neutrosophic beta distribution with parameters \code{shape1}=\eqn{\alpha_N} and
\code{shape2}=\eqn{\beta_N} has density
\deqn{f_N(X)=\frac{1}{B\left(\alpha_N, \beta_N\right)}X^{\alpha_N-1}(1-X)^{\beta_N-1}}
for  \eqn{\alpha_N > 0}, the first shape parameter, and \eqn{\beta_N > 0},
the second shape parameter and \eqn{0 \le x \le 1}. The function \eqn{B(a,b)}
return the beta function that can be calculated by \code{\link{beta}}.
}
\examples{
x <- seq(0.1, 1, length.out = 21) #'
pnbd(x, alpha = 2, beta = 1)

x2 <- matrix(seq(0.1, 1, length.out = 40), ncol = 2)
pnbd(x2, alpha = c(1, 2), beta = c(2, 2))

dnbd(x, alpha = 1, beta = 2)

dnbd(x2, alpha = c(1, 2), beta = c(2, 2))


qnbd(x, alpha = 1, beta = 2)
qnbd(x2, alpha = c(1, 2), beta = c(2, 2))

n <- 10
rnbd(n, alpha = 2, beta = 1)
rnbd(n, alpha = c(1, 2), beta = c(1, 1))
}
\references{
Sherwani, R. Ah. K., Naeem, M., Aslam, M., Reza, M. A., Abid, M., Abbas, S. (2021).
    Neutrosophic beta distribution with properties and applications. Neutrosophic Sets and Systems, 41, 209-214.
}
