% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbind.R
\name{NBIND}
\alias{NBIND}
\alias{dbind}
\alias{pbind}
\alias{qbind}
\alias{rbind}
\title{Neutrosophic Binomial Distribution (NBIND)}
\usage{
dbind(x, size, prob)

pbind(q, size, prob, lower.tail = TRUE)

qbind(p, size, prob)

rbind(n, size, prob)
}
\arguments{
\item{x}{a vector or matrix of observations for which the pdf needs to be computed.}

\item{size}{number of trials (zero or more), which must be a positive interval.}

\item{prob}{probability of success on each trial.}

\item{q}{a vector or matrix of quantiles for which the cdf needs to be computed.}

\item{lower.tail}{logical; if TRUE (default), probabilities are
\eqn{P(X \ge x)}; otherwise, \eqn{P(X >x)}.}

\item{p}{a vector or matrix of probabilities for which the quantile needs to be computed.}

\item{n}{number of random values to be generated.}
}
\value{
\code{pbind} gives the distribution function,
 \code{dbind} gives the density,
 \code{qbind} gives the quantile function and
 \code{rbind} generates random variables from the neutrosophic Poisson Distribution.
}
\description{
Density, distribution function, quantile function and random
generation for the nuetrosophic Binomial distribution with
parameters \eqn{n} and \eqn{p_N}.
}
\details{
The neutrosophic Poisson distribution with parameter \eqn{\lambda_N}
has the density
\deqn{f_X(x)=\bigg(\begin{array}{c}n \\ x\end{array}\bigg) p_N^{x}\left(1-p_N\right)^{n-x}}
for \eqn{n \in \{0, 1, 2, \ldots\}} and \eqn{p_N \in (pL, \lambda_U)} which must be \eqn{0<p_N<1}
and \eqn{x \in \{0, 1, 2, \ldots\}}.
}
\examples{
dbind(x, size = 2, prob= 0.5)
dbind(x2, lambda = c(2, 2))
x <- 1:10
x2 <- matrix(1:20, ncol = 2)
pbind(x, size = 2, prob= 0.5)
pbind(x2, size = c(2,2), prob = c(.3 ,.6))
q1 <- seq(0.1, 1, length.out = 40)
qbind(q1, size = 2, prob= 0.5)
q2 <- matrix(seq(0.1, 1, length.out = 40), ncol = 2)
qbind(q2, lambda = c(2, 2))
n <- 10
rbind(n, size = 2, prob= 0.5)
rbind(n, lambda = c(1, 2))
}
\references{
Granados, C. (2022).
       Some discrete neutrosophic distributions with neutrosophic parameters based on neutrosophic random variables.
        \emph{Hacettepe Journal of Mathematics and Statistics}, 51(5), 1442-1457.
}
