% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ned.R
\name{NED}
\alias{NED}
\alias{pned}
\alias{dned}
\alias{qned}
\alias{rned}
\title{Neutrosophic Exponential Distribution (NED)}
\usage{
pned(q, theta)

dned(x, theta)

qned(p, theta)

rned(n, theta)
}
\arguments{
\item{theta}{the positive value or vector lower and upper of the first shape parameter.}

\item{x, q}{vector or matrix lower and upper of quantiles at which the pdf or cdf needs to be computed.}

\item{p}{vector or matrix lower and upper of probabilities at which the quantile needs to be computed.}

\item{n}{number of random numbers to be generated.}
}
\value{
\code{pned} gives the distribution function,
 \code{dned} gives the density,
 \code{qned} gives the quantile function and
 \code{rned} generates random variables from the neutrosophic exponential distribution.
}
\description{
Density, distribution function, quantile function and random generation for
the nuetrosophic exponential distribution with parameter \code{rate}=\eqn{\theta_N}.
}
\details{
The neutrosophic exponential distribution with parameters \code{rate}=\eqn{\theta_N}
has density
\deqn{f_N(x)=\theta_N \exp \left(-x \theta_N\right)}
for \eqn{x \ge 0} and \eqn{\theta_N > 0}, the rate parameter.
}
\examples{
x <- seq(0.01, 1, length.out = 21)
pned(x, theta = 1)

x2 <- matrix(seq(0.01, 2, length.out = 40), ncol = 2)
pned(x2, theta = c(2, 3))

dned(x, theta = 2)

dned(x2, theta = c(2, 2))

qned(x, theta = 2)

qned(x2, theta = c(2, 2))

n <- 10
rned(n, theta = 1)

rned(n, theta = c(1, 2))
}
\references{
Duan, W., Q., Khan, Z., Gulistan, M., Khurshid, A. (2021). Neutrosophic
Exponential Distribution: Modeling and Applications for Complex Data Analysis,
\emph{Complexity}, 2021, 1-8.
}
