% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nspois.R
\name{Neutrosophic Poisson}
\alias{Neutrosophic Poisson}
\alias{dnspois}
\alias{pnspois}
\alias{qnspois}
\alias{rnspois}
\title{Neutrosophic Poisson Distribution}
\usage{
dnspois(x, lambda)

pnspois(q, lambda, lower.tail = TRUE)

qnspois(p, lambda)

rnspois(n, lambda)
}
\arguments{
\item{x}{a vector or matrix of observations for which the pdf needs to be computed.}

\item{lambda}{the mean, which must be a positive interval.}

\item{q}{a vector or matrix of quantiles for which the cdf needs to be computed.}

\item{lower.tail}{logical; if TRUE (default), probabilities are
\eqn{P(X \ge x)}; otherwise, \eqn{P(X >x)}.}

\item{p}{a vector or matrix of probabilities for which the quantile needs to be computed.}

\item{n}{number of random values to be generated.}
}
\value{
\code{pnspois} gives the distribution function,
 \code{dnspois} gives the density,
 \code{qnspois} gives the quantile function and
 \code{rnspois} generates random variables from the neutrosophic Poisson Distribution.
}
\description{
Density, distribution function, quantile function and random
generation for the nuetrosophic Poisson distribution with
parameter \eqn{\lambda_N}.
}
\details{
The neutrosophic Poisson distribution with parameter \eqn{\lambda_N}
has the density
\deqn{f_N(x)= \exp\{-\lambda_N\}  \frac{\left(\lambda_N\right)^x}{x !}}
for \eqn{\lambda_N \in (\lambda_L, \lambda_U)} which must be a positive
interval and \eqn{x \in \{0, 1, 2, \ldots\}}.
}
\examples{
dnspois(x = 1, lambda = 2)
pnspois(q = 1, lambda = 1)
qnspois(p = c(0.25,0.5,0.75), lambda = 2)
rnspois(n = 10, lambda = 1)
}
\references{
Alhabib, R., Ranna, M. M., Farah, H., Salama, A. A. (2018).
       Some neutrosophic probability distributions.
       \emph{Neutrosophic Sets and Systems},  22, 30-38.
}
