% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngd.R
\name{NGD}
\alias{NGD}
\alias{dngd}
\alias{pngd}
\alias{qngd}
\alias{rngd}
\title{Neutrosophic Gamma Distribution (NGD)}
\usage{
dngd(x, alpha, lambda)

pngd(q, alpha, lambda, lower.tail = TRUE)

qngd(p, alpha, lambda)

rngd(n, alpha, lambda)
}
\arguments{
\item{x}{a vector or matrix of observations for which the pdf needs to be computed.}

\item{alpha}{the shape parameter, which must be a positive interval.}

\item{lambda}{the scale parameter, which must be a positive interval.}

\item{q}{a vector or matrix of quantiles for which the cdf needs to be computed.}

\item{lower.tail}{logical; if TRUE (default), probabilities are
\eqn{P(X \ge x)}; otherwise, \eqn{P(X >x)}.}

\item{p}{a vector or matrix of probabilities for which the quantile needs to be computed.}

\item{n}{number of random values to be generated.}
}
\value{
\code{pngd} gives the distribution function,
 \code{dngd} gives the density,
 \code{qngd} gives the quantile function and
 \code{rngd} generates random variables from the neutrosophic gamma distribution.
}
\description{
Density, distribution function, quantile function and random generation for
the nuetrosophic gamma distribution with parameter \code{shape}=\eqn{\alpha_N}
and \code{scale}=\eqn{\lambda_N}.
}
\details{
The neutrosophic gamma distribution with parameters \eqn{\alpha_N} and
\eqn{\lambda_N} has density
\deqn{f_n(x)=\frac{1}{\Gamma(\alpha_n) \lambda_n^{\alpha_n}} x^{\alpha_n-1} \exp\{-\left(x / \lambda_n\right)\}}
for \eqn{x \ge 0}, \eqn{\alpha_N \in (\alpha_L, \alpha_U)}, the shape
parameter which must be a positive interval and
\eqn{\lambda_n \in (\lambda_L, \lambda_U)}, the scale parameter which
must be a positive interval. Here, \eqn{\Gamma(\cdot)} is gamma
function implemented by \code{\link{gamma}}.
}
\examples{

dngd(x = 0.1, alpha = c(2,3), lambda = c(1,2))

x <- matrix(c(1, 1, 2, 2.2, 3, 3.5), ncol = 2, byrow = TRUE)
dngd(x, alpha = c(1, 2), lambda = c(2, 2))
pngd(q = c(0.1, 0.1), alpha = c(1,1.5), lambda = c(2,2))


qngd(p = 0.1, alpha = c(1, 2), lambda = c(2, 2))
Simulate 10 numbers
rngd(n=10, alpha = c(1, 2), lambda = c(1, 1))
}
\references{
Khan, Z., Al-Bossly, A., Almazah, M. M. A., and Alduais, F. S. (2021).
   On statistical development of neutrosophic gamma distribution with
   applications to complex data analysis, \emph{Complexity}, 2021, Article ID 3701236.
}
