% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nge.R
\name{NGED}
\alias{NGED}
\alias{pnged}
\alias{dnged}
\alias{qnged}
\alias{rnged}
\alias{hnged}
\alias{snged}
\title{Neutrosophic Generalized Exponential Distribution (NGED)}
\usage{
pnged(x, v = 1, delta = 2)

dnged(x, v = 1, delta = 2)

qnged(q, v = 1, delta = 2)

rnged(n, v = 1, delta = 2)

hnged(x, v = 1, delta = 2)

snged(x, v = 1, delta = 2)
}
\arguments{
\item{x}{scaler or vector of values at which the pdf or cdf needs to be computed.}

\item{v}{the value of the first shape parameter.}

\item{delta}{the value of the second shape parameter, must be positive, the default is 2.}

\item{q}{scaler or vector of probabilities at which the quantile needs to be computed.}

\item{n}{number of random numbers to be generated.}
}
\value{
\code{pnged} gives the distribution function,
 \code{dnged} gives the density,
 \code{qnged} gives the quantile function,
 \code{hnged} gives the hazard function,
 \code{snged} gives the survival function and
 \code{rnged} generates random variables from the Neutrosophic Generalized Exponential (NGE).
}
\description{
Computes the pdf, cdf, hdf, quantile and random numbers of the
}
\examples{
x <- seq(0, 1, length.out = 21)
x2 <- matrix(seq(0, 2, length.out = 40),ncol=2)
pnged(x)
pnged(x, v = 2, delta = 1)
dnged(x, v = 1, delta = 2)
curve(dnged, -3, 3)
qnged(x, v = 1, delta = 3)
n <- 10
rnged(n, v = 2, delta=1)
hnged(x, v = 1, delta = 3)
curve(hnged, 0, 3)
snged(x, v = 1, delta = 3)
curve(snged, 0, 3)
snged(x2, v = c(1,2), delta = c(2,2))
}
\references{
Rao, Gadde Srinivasa, Mina Norouzirad, and Danial Mazarei. "Neutrosophic Generalized Exponential Distribution with Application." Neutrosophic Sets and Systems 55.1 (2023): 28.
}
