% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nged.R
\name{NGED}
\alias{NGED}
\alias{pnged}
\alias{dnged}
\alias{qnged}
\alias{rnged}
\title{Neutrosophic Generalized Exponential Distribution (NGED)}
\usage{
pnged(q, nu, delta)

dnged(x, nu, delta)

qnged(p, nu, delta)

rnged(n, nu = 1, delta = 2)
}
\arguments{
\item{nu}{scale parameter, must be positive.}

\item{delta}{shape parameter, must be positive.}

\item{x, q}{quantiles at which the pdf or cdf needs to be computed.}

\item{p}{probabilities at which the quantile needs to be computed.}

\item{n}{number of random numbers to be generated.}
}
\value{
\code{pnged} gives the distribution function,
 \code{dnged} gives the density,
 \code{qnged} gives the quantile function and
 \code{rnged} generates random variables
 from the neutrosophic generalized exponential distribution.
}
\description{
Density, distribution function, quantile function and random generation for
the nuetrosophic generalized exponential distribution with shape parameter
\eqn{\delta} and scale parameter \eqn{\nu}.
}
\details{
The neutrosophic generalized exponential distribution with parameters
\code{shape}=\eqn{\delta} and \code{scale}=\eqn{\nu} has density
\deqn{f_n(x)=\frac{\delta_N}{\nu_N}\left(1-\exp \left\{-\frac{x_N}{\nu_N}\right\}\right)^{\delta_N-1} e^{\left\{-\frac{x_N}{\nu_N}\right\}}}
for\eqn{x \ge 0}, \eqn{\delta_N > 0}, the shape parameter,
and \eqn{\nu_N > 0}, the scale parameter.
}
\examples{
x <- seq(0.1, 1, length.out = 21)
pnged(x, nu = 2, delta = 1)

x2 <- matrix(seq(0.1, 1, length.out = 40), ncol = 2)
pnged(x2, nu = c(1, 2), delta = c(2, 2))
dnged(x, nu = 1, delta = 2)

dnged(x2, nu = c(1, 2), delta = c(2, 2))
qnged(x, nu = 1, delta = 2)

qnged(x2, nu = c(1, 2), delta = c(2, 2))
n <- 10
rnged(n, nu = 2, delta = 1)
rnged(n, nu = c(1, 2), delta = c(1, 1))
}
\references{
Rao, G. S., Norouzirad, M., and Mazarei . D. (2023). Neutrosophic
   Generalized Exponential Distribution with Application.
   \emph{Neutrosophic Sets and Systems}, 55, 471-485.
}
