% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnd.R
\name{NND}
\alias{NND}
\alias{pnnd}
\alias{dnnd}
\alias{qnnd}
\alias{rnnd}
\title{Neutrosophic Normal Distribution (NND)}
\usage{
pnnd(x, mu = 1, sigma = 2)

dnnd(x, mu = 1, sigma = 2)

qnnd(q, mu = 1, sigma = 2)

rnnd(n, mu = 1, sigma = 2)
}
\arguments{
\item{x}{scaler or vector or matrix lower and upper of values at which the pdf or cdf needs to be computed.}

\item{mu}{the value or vector lower and upper of the first parameter.}

\item{sigma}{the value or vector lower and upper of the second parameter, must be positive.}

\item{q}{scaler or vector of probabilities at which the quantile needs to be computed.}

\item{n}{number of random numbers to be generated.}
}
\value{
\code{pnnd} gives the distribution function,
 \code{dnnd} gives the density,
 \code{qnnd} gives the quantile function and
 \code{rnnd} generates random variables from the Neutrosophic Normal Distribution (NND).
}
\description{
Computes the pdf, cdf, quantile and random numbers of the nuetrosophic generalized exponential distribution.
\deqn{X_N \sim N_N\left(\mu_N, \sigma_N\right)=\frac{1}{\sigma_N \sqrt{2 \pi}} e^{\left(\frac{\left(X-\mu_N\right)^2}{2 \sigma_N^2}\right)}}
for   \eqn{\mu_N}, the first parameter, and \eqn{\sigma_N > 0}, the second parameter.
}
\examples{
x <- seq(0.1, 1, length.out = 21)
x2 <- matrix(seq(0.1, 2, length.out = 40), ncol = 2)
pnnd(x)
pnnd(x, mu = 2, sigma = 1)
pnnd(x2, mu = c(1, 2), sigma = c(2, 2))
dnnd(x, mu = 1, sigma = 2)
curve(dnnd, .1, 3)
dnnd(x2, mu = c(1, 2), sigma = c(2, 2))
qnnd(x, mu = 1, sigma = 2)
x2 <- matrix(seq(0.1, 1, length.out = 40), ncol = 2)
qnnd(x2, mu = c(1, 2), sigma = c(2, 2))
n <- 100
rnnd(n, mu = 2, sigma = 1)
rnnd(n, mu = c(1, 2), sigma = c(1, 1))
}
\references{
Patro, S. and F. Smarandache, The Neutrosophic Statistical Distribution, More Problems, More Solutions. 2016: Infinite Study.
}
