% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnd.R
\name{NND}
\alias{NND}
\alias{pnnd}
\alias{dnnd}
\alias{qnnd}
\alias{rnnd}
\title{Neutrosophic Normal Distribution (NND)}
\usage{
pnnd(q, mu, sigma)

dnnd(x, mu, sigma)

qnnd(p, mu, sigma)

rnnd(n, mu, sigma)
}
\arguments{
\item{mu}{the value or vector lower and upper of the first parameter.}

\item{sigma}{the positive value or vector lower and upper of the second parameter.}

\item{x, q}{vector or matrix lower and upper of quantiles at which the pdf or cdf needs to be computed.}

\item{p}{vector or matrix lower and upper of probabilities at which the quantile needs to be computed.}

\item{n}{number of random numbers to be generated.}
}
\value{
\code{pnnd} gives the distribution function,
 \code{dnnd} gives the density,
 \code{qnnd} gives the quantile function and
 \code{rnnd} generates random variables from the neutrosophic normal distribution.
}
\description{
Density, distribution function, quantile function and random generation for
the nuetrosophic generalized exponential distribution with parameters mean,
\eqn{\mu}, and standard deviation \eqn{\sigma}.
}
\details{
The neutrosophic normal distribution with parameters with \code{mean}=\eqn{\mu_N}
and \code{std}=\eqn{\sigma_N} has density
\deqn{f_N(x) = \frac{1}{\sigma_N \sqrt{2 \pi}} e^{\left(\frac{\left(X-\mu_N\right)^2}{2 \sigma_N^2}\right)}}
for \eqn{-\infty < x < \infty}, \eqn{\mu_N}, the mean, and \eqn{\sigma_N > 0},
the standard deviation.
}
\examples{
x <- seq(0.1, 1, length.out = 21)
pnnd(x, mu = c(2, 2), sigma = c(1, 1))

x2 <- matrix(seq(0.1, 1, length.out = 40), ncol = 2)
pnnd(x2, mu = c(1, 2), sigma = c(2, 2))
dnnd(x = 0.5, mu = 1, sigma = 2)

x1 <- c(-0.8, 0.2, 1.6, 3.9)
dnnd(x = 0.5, mu = c(1, 1), sigma = c(2, 2))

x2 <- matrix(seq(-3, 3, length.out = 10), nrow = 2, ncol = 5)
dnnd(x2, mu = c(1, 2), sigma = c(2, 2))
q1 <- c(0.01)
qnnd(q1, mu = 1, sigma = 2)
x2 <- matrix(seq(0.1, 1, length.out = 40), ncol = 2)
qnnd(x2, mu = c(1, 2), sigma = c(2, 2))
n <- 100
rnnd(n, mu = c(1, 1), sigma = c(2, 2))
rnnd(n, mu = c(1, 2), sigma = c(1, 1))
}
\references{
Patro, S. and F. Smarandache, The Neutrosophic Statistical Distribution, More Problems, More Solutions. 2016: Infinite Study.
}
