% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npd.R
\name{NPD}
\alias{NPD}
\alias{pnpd}
\alias{dnpd}
\alias{qnpd}
\alias{rnpd}
\title{Neutrosophic Poisson Distribution (NPD)}
\usage{
pnpd(x, lambda = 2)

dnpd(x, lambda = 2)

qnpd(q, lambda = 2)

rnpd(n, lambda = 2)
}
\arguments{
\item{x}{scaler or vector or matrix lower and upper of values (non-negative integer) at which the pdf or cdf needs to be computed.}

\item{lambda}{the value or vector lower and upper of (non-negative)the first shape parameter, must be positive.}

\item{q}{scaler or vector of probabilities at which the quantile needs to be computed.}

\item{n}{number of random numbers to be generated.}
}
\value{
\code{pnpd} gives the distribution function,
 \code{dnpd} gives the density,
 \code{qnpd} gives the quantile function and
 \code{rnpd} generates random variables from the Neutrosophic Poisson Distribution (NPD).
}
\description{
Computes the pdf, cdf, hdf, quantile and random numbers of the
\deqn{f_N(x)=e^{-\lambda_N} \cdot \frac{\left(\lambda_N\right)^x}{x !}}
for   \eqn{\lambda_N > 0}, the first shape parameter.
}
\examples{
x <- 1:10
x2 <- matrix(1:20, ncol = 2)
pnpd(x)
pnpd(x, lambda = 1)
pnpd(x2, lambda = c(2, 3))
dnpd(x, lambda = 2)
curve(dnpd, .1, 3)
dnpd(x2, lambda = c(2, 2))
x3 <- seq(0.1, 1, length.out = 40)
qnpd(x3, lambda = 2)
x2 <- matrix(seq(0.1, 1, length.out = 40), ncol = 2)
qnpd(x2, lambda = c(2, 2))
n <- 10
rnpd(n, lambda = 1)
rnpd(n, lambda = c(1, 2))
}
\references{
Alhabib, R., et al., Some neutrosophic probability distributions. Neutrosophic Sets and Systems, 2018. 22: p.30-38.
}
