% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrd.R
\name{NRD}
\alias{NRD}
\alias{pnrd}
\alias{dnrd}
\alias{qnrd}
\alias{rnrd}
\title{Neutrosophic Rayleigh Distribution (NRD)}
\usage{
pnrd(q, theta)

dnrd(x, theta)

qnrd(p, theta)

rnrd(n, theta)
}
\arguments{
\item{theta}{the value or vector lower and upper of the first shape parameter, must be positive.}

\item{x, q}{vector or matrix lower and upper of quantiles at which the pdf or cdf needs to be computed.}

\item{p}{vector or matrix lower and upper of probabilities at which the quantile needs to be computed.}

\item{n}{number of random numbers to be generated.}
}
\value{
\code{pnrd} gives the distribution function,
 \code{dnrd} gives the density,
 \code{qnrd} gives the quantile function and
 \code{rnrd} generates random variables from the Neutrosophic Rayleigh Distribution (NRD).
}
\description{
Density, distribution function, quantile function and random generation for
the nuetrosophic Rayleigh distribution with parameter \eqn{\theta_N}.
}
\details{
The neutrosophic Rayleigh distribution with parameter \eqn{\theta} has the density
\deqn{f_N(x)=\frac{x}{\theta_N^2} e^{-\frac{1}{2}\left(\frac{x}{\theta_N}\right)^2}}
for  \eqn{\theta_N > 0}.
}
\examples{
x <- seq(0.01, 1, length.out = 21)
pnrd(x, theta = 1)

x2 <- matrix(seq(0.01, 1, length.out = 40), ncol = 2) #'
pnrd(x2, theta = c(2, 3))
dnrd(x, theta = 2)
dnrd(x2, theta = c(2, 2))
qnrd(x, theta = 2)

qnrd(x2, theta = c(2, 2))
n <- 10
rnrd(n, theta = 1)
rnrd(n, theta = c(1, 2))
}
\references{
Khan, Z., Gulistan, M., Kausar, N. and Park, C. (2021). Neutrosophic Rayleigh Model With Some Basic Characteristics and Engineering Applications, in \emph{IEEE Access}, 9, 71277-71283.
}
