% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrd.R
\name{NRD}
\alias{NRD}
\alias{pnrd}
\alias{dnrd}
\alias{qnrd}
\alias{rnrd}
\title{Neutrosophic Rayleigh Distribution (NRD)}
\usage{
pnrd(x, theta = 2)

dnrd(x, theta = 2)

qnrd(q, theta = 2)

rnrd(n, theta = 2)
}
\arguments{
\item{x}{scaler or vector or matrix lower and upper of values at which the pdf or cdf needs to be computed.}

\item{theta}{the value or vector lower and upper of the first shape parameter.}

\item{q}{scaler or vector of probabilities at which the quantile needs to be computed.}

\item{n}{number of random numbers to be generated.}
}
\value{
\code{pnrd} gives the distribution function,
 \code{dnrd} gives the density,
 \code{qnrd} gives the quantile function and
 \code{rnrd} generates random variables from the Neutrosophic Rayleigh (NGE).
}
\description{
Computes the pdf, cdf, hdf, quantile and random numbers of the
}
\examples{
x <- seq(0.01, 1, length.out = 21)
x2 <- matrix(seq(0.01, 2, length.out = 40), ncol = 2)
pnrd(x)
pnrd(x, theta = 1)
pnrd(x2, theta = c(2, 3))
dnrd(x, theta = 2)
curve(dnrd, .1, 3)
dnrd(x2, theta = c(2, 2))
qnrd(x, theta = 2)
x2 <- matrix(seq(0.1, 1, length.out = 40), ncol = 2)
qnrd(x2, theta = c(2, 2))
n <- 10
rnrd(n, theta = 1)
rnrd(n, theta = c(1, 2))
}
\references{
Khan, Zahid, et al. "Neutrosophic Rayleigh model with some basic characteristics and engineering applications." IEEE Access 9 (2021): 71277-71283.
}
