% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwd.R
\name{NWD}
\alias{NWD}
\alias{pnwd}
\alias{dnwd}
\alias{qnwd}
\alias{rnwd}
\title{Neutrosophic Weibull Distribution (NWD)}
\usage{
pnwd(q, beta, alpha)

dnwd(x, beta, alpha)

qnwd(p, beta, alpha)

rnwd(n, beta, alpha)
}
\arguments{
\item{beta}{the value or vector lower and upper of the first shape parameter. must be positive.}

\item{alpha}{the value or vector lower and upper of the second scale parameter, must be positive.}

\item{x, q}{vector or matrix lower and upper of quantiles at which the pdf or cdf needs to be computed.}

\item{p}{vector or matrix lower and upper of probabilities at which the quantile needs to be computed.}

\item{n}{number of random numbers to be generated.}
}
\value{
\code{pnwd} gives the distribution function,
 \code{dnwd} gives the density,
 \code{qnwd} gives the quantile function and
 \code{rnwd} generates random variables from the Neutrosophic Weibull Distribution (NWD).
}
\description{
Density, distribution function, quantile function and random generation for
the nuetrosophic Weibull distribution with scale parameter \eqn{\alpha_N} and
shape parameter \eqn{\beta_N}.
}
\details{
The neutrosophic Rayleigh distribution with parameters \eqn{\alpha_N} and
\eqn{\beta_N} has the density
\deqn{f_N(x)=\frac{\beta_N}{\alpha_N^{\beta_N}} x^{\beta_N-1}
    e^{-\left(x / \alpha_N\right)^{\beta_N}}}
for \eqn{x > 0}, \eqn{\beta_N > 0}, shape parameter, and \eqn{\alpha_N > 0},
scale parameter.
}
\examples{
p <- seq(0.1, 1, length.out = 21)
p2 <- matrix(seq(0.1, 2, length.out = 40), ncol = 2)
pnwd(p, beta = 1, alpha = 2)
pnwd(p2, beta = c(1, 2), alpha = c(2, 2))
dnwd(x, beta = 1, alpha = 2)

dnwd(x2, beta = c(1, 2), alpha = c(2, 2))
qnwd(x, beta = 1, alpha = 2)
x2 <- matrix(seq(0.1, 1, length.out = 40), ncol = 2)
qnwd(x2, beta = c(1, 2), alpha = c(2, 2))
n <- 10
rnwd(n, beta = 1, alpha = 2)
rnwd(n, beta = c(1, 2), alpha = c(1, 1))
}
\references{
Alhasan, K. F. H. and Smarandache, F. (2019). Neutrosophic Weibull
   distribution and Neutrosophic Family Weibull Distribution,
   \emph{Neutrosophic Sets and Systems}, 28, 191-199.
}
