% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwd.R
\name{NWD}
\alias{NWD}
\alias{dnwd}
\alias{qnwd}
\alias{rnwd}
\title{Neutrosophic Weibull Distribution (NWD)}
\usage{
dnwd(x, beta, alpha)

qnwd(p, beta, alpha)

rnwd(n, beta, alpha)
}
\arguments{
\item{x}{a vector or matrix of observations for which the pdf needs to be computed.}

\item{beta}{shape parameter, which must be a positive interval.}

\item{alpha}{scale parameter, which must be a positive interval.}

\item{p}{a vector or matrix of probabilities for which the quantile needs to be computed.}

\item{n}{number of random values to be generated.}

\item{q}{a vector or matrix of quantiles for which the cdf needs to be computed.}

\item{lower.tail}{logical; if TRUE (default), probabilities are
\eqn{P(X \ge x)}; otherwise, \eqn{P(X >x)}.}
}
\value{
\code{pnwd} gives the distribution function,
 \code{dnwd} gives the density,
 \code{qnwd} gives the quantile function and
 \code{rnwd} generates random variables from the neutrosophic Weibull dDistribution.
}
\description{
Density, distribution function, quantile function and random
generation for the nuetrosophic Weibull distribution with scale
parameter \eqn{\alpha_N} and shape parameter \eqn{\beta_N}.
}
\details{
The neutrosophic Rayleigh distribution with parameters
\eqn{\alpha_N} and \eqn{\beta_N} has the density
\deqn{f_N(x)=\frac{\beta_N}{\alpha_N^{\beta_N}} x^{\beta_N-1}
    \exp\{-\left(x / \alpha_N\right)^{\beta_N}\}}
for \eqn{\beta_N \in (\beta_L, \beta_U)} the shape parameter must
be a positive interval, \eqn{\alpha_N \in (\alpha_L,\alpha_U)},
the scale parameter which be a positive interval, and \eqn{x > 0}.
}
\examples{
dnwd(x, beta = 1, alpha = 2)

dnwd(x2, beta = c(1, 2), alpha = c(2, 2))
qnwd(x, beta = 1, alpha = 2)
x2 <- matrix(seq(0.1, 1, length.out = 40), ncol = 2)
qnwd(x2, beta = c(1, 2), alpha = c(2, 2))
n <- 10
rnwd(n, beta = 1, alpha = 2)
rnwd(n, beta = c(1, 2), alpha = c(1, 1))
}
\references{
Alhasan, K. F. H. and Smarandache, F. (2019). Neutrosophic Weibull
   distribution and Neutrosophic Family Weibull Distribution,
   \emph{Neutrosophic Sets and Systems}, 28, 191-199.
}
